checkers/requirements.txt
checklib
beautifulsoup4==4.11.1
python-mimeparse==1.6.0